CARDS-1375: Add an option to block advancing in a form unless all required fields have been filled out * Renamed the new FormUtility to FormUtilities * Moved the utility function `hasWarningFlags` from `AnswerInstructions` to FormUtilities * Modified hasWarningFlags   * to accept either a [UUID, dataObject] pair or directly a data object as a parameter   * to also return true for the DRAFT flag * Modified the criteria for requireCompletion to use hasWarningFlags instead of looking only for the INCOMPLETE flag
