+++ modules/data-entry/src/main/frontend/src/questionnaire/AnswerInstructions.jsx
@@ -26,10 +26,7 @@ import withStyles from '@mui/styles/withStyles';
 
 import QuestionnaireStyle from "./QuestionnaireStyle";
 
-// Check if an answer has been automatically flagged as incomplete or invalid
-export function hasWarningFlags (data) {
-  return !!(data?.[1]?.statusFlags?.includes('INCOMPLETE') || data?.[1]?.statusFlags?.includes('INVALID'))
-}
+import { hasWarningFlags } from "./FormUtilities";
 
 // Display instructions regarding how many answers must be provided to a question,
 // based on minAnswers and maxAnswers from the question definition
