+++ modules/data-entry/src/main/frontend/src/questionnaire/FormUtilities.jsx
@@ -17,6 +17,17 @@
 //  under the License.
 //
 
+// Check if an answer has been automatically flagged as incomplete or invalid
+// The data parameter can come either as a [key, value] array or as an object
+export function hasWarningFlags (data) {
+    return (
+      Array.isArray(data) ?
+        hasWarningFlags(data?.[1])
+      :
+        !!(data?.statusFlags?.some(f => ["INCOMPLETE", "INVALID", "DRAFT"].includes(f)))
+    );
+}
+
 // A helper recursive function to loop through the sections/questions data 
 // and collect all questions that have incomplete status among it's statusFlags
 // Recursively build a map of uuid->path of all incomplete questions
@@ -24,10 +35,10 @@ export function getIncompleteQuestionsMap (sectionJson) {
     let retFields = {};
     Object.entries(sectionJson).map(([title, object]) => {
         // We only care about children that are cards/Answers or cards:AnswerSections
-        if (object["sling:resourceSuperType"] == "cards/Answer" && object.statusFlags?.includes("INCOMPLETE")) {
+        if (object["sling:resourceSuperType"] == "cards/Answer" && hasWarningFlags(object)) {
           // If this is an cards:Question, we copy the entire path to the array
           retFields[object.question["jcr:uuid"]] = object.question["@path"];
-        } else if (object["jcr:primaryType"] == "cards:AnswerSection" && object.statusFlags?.includes("INCOMPLETE")) {
+        } else if (object["jcr:primaryType"] == "cards:AnswerSection" && hasWarningFlags(object)) {
           // if this is matrix - save the section path as this section is rendered in Question component
           if ("matrix" === object.section["displayMode"]) {
             retFields[object.question["jcr:uuid"]] = object.question["@path"];
@@ -65,14 +76,14 @@ export function parseSectionOrQuestionnaire (sectionJson) {
 // Gets the first visible incomplete question element from the form json
 export function getFirstIncompleteQuestionEl (json) {
     //if the form is incomplete itself
-    if (json?.statusFlags?.includes("INCOMPLETE")) {
+    if (hasWarningFlags(json)) {
       // Get an array of the question uuids in the order defined by questionnaire
       let allUuids = parseSectionOrQuestionnaire(json.questionnaire);
       // Get the map of uuid:path of the incomplete answers
       let incompleteQuestions = getIncompleteQuestionsMap(json);
 
       if (Object.keys(incompleteQuestions).length > 0) {
-	    // Loop through the question uuids in the order defined by questionnaire
+	// Loop through the question uuids in the order defined by questionnaire
         for (const uuid of allUuids) {
           let firstEl = document.getElementById(incompleteQuestions[uuid]);
           // Find first visible element by path from incomplete answers map
@@ -81,7 +92,6 @@ export function getFirstIncompleteQuestionEl (json) {
           }
         }
       }
-
     }
     return null;
 }
